TX2_MobileFaceNet/requirements.txt
absl-py==0.8.1
astor==0.8.0
certifi==2019.9.11
dlib==19.18.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.1
h5py==2.10.0
imutils==0.5.3
#Keras-Applications==1.0.8
#Keras-Preprocessing==1.1.0
Markdown==3.1.1
mock==3.0.5
mtcnn==0.0.9
numpy==1.17.3
onnx==1.6.0
onnx-tf==1.3.0
#onnxruntime==0.5.0
#opencv-python==4.1.1.26
opt-einsum==3.1.0
protobuf==3.10.0
PyYAML==5.1.2
six==1.12.0
#tensorboard==1.13.1
#tensorflow==1.13.1
#tensorflow-estimator==1.13.0
termcolor==1.1.0
typing-extensions==3.7.4
Werkzeug==0.16.0
wincertstore==0.2
wrapt==1.11.2
